thêm lấy danh sách khách hàng quảng cáo

+++ WindowsFormsApp2/DAO/HangDAO.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -23,13 +24,15 @@ namespace DAO
             DataTable dt = this.dp.ExecuteQuery(query);
 
             List<HangDTO> hangS = new List<HangDTO>();
+            
             foreach (DataRow dr in dt.Rows)
             {
                 HangDTO hang = new HangDTO((int)dr["MAHANG"], (int)dr["MANVPHUTRACH"], (string)dr["TENHANG"],
-                    (int)dr["SOLUONGCONLAI"], (float)dr["DONGIA"]);
+                    (int)dr["SOLUONGCONLAI"], (double)dr["DONGIA"]);
                 
                 hangS.Add(hang);
-            }
+            }                
+                   
             return hangS;
         }
     }
